[refactor] 마이페이지 스타일드 컴포넌트 분리 ,비지니스로직 분리
